requirements.txt
streamlit==1.32.0
langchain==0.1.9
langchain-groq==0.0.5
groq==0.4.1
python-dotenv==1.0.1
PyPDF2==3.0.1
google-generativeai==0.3.2
numpy==1.24.32.0.1
python-dotenv==1.0.1
langchain==0.1.9
langchain-groq
groq==0.4.1
PyPDF2==3.0.1
google-generativeai==0.3.2
